requirements.txt
ultralytics>=8.3.0,<9.0.0
opencv-python>=4.10.0,<5.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
python-multipart>=0.0.12,<1.0.0
pillow>=11.0.0,<12.0.0
groq>=0.13.0,<1.0.0
google-genai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.32.0,<3.0.0
